[#26463] YSQL: Fix integer overflow in seek next estimation for inner table in BNL
Summary:
Due to a bug in the cost model, sometimes the seeks and nexts estimation could
be negative. This would result in negative cost estimation. This was caused
by an integer overflow. This was fixed by using double instead of int.
Jira: DB-15827

Test Plan: ./yb_build.sh --java-test 'org.yb.pgsql.TestPgCostModelSeekNextEstimation'

Reviewers: mihnea, mtakahara

Reviewed By: mtakahara

Subscribers: svc_phabricator, kannan, yql

Differential Revision: https://phorge.dev.yugabyte.com/D42672

+++ src/postgres/src/backend/utils/adt/selfuncs.c
@@ -6680,7 +6680,7 @@ yb_batch_expr_size(PlannerInfo *root, Index path_relid, Node *batched_expr)
 
 	root->yb_cur_batched_relids = NULL;
 
-	int			num_outer_tuples = get_loop_count(root, path_relid,
+	double		num_outer_tuples = get_loop_count(root, path_relid,
 												  other_varnos);
 
 	root->yb_cur_batched_relids = batched_relids;
